output to $HOME instead of /home/$USER
if you run this from a root prompt, or as any other user whose home directory does not live at /home/$USER, it fails with the message "/home/root does not exist".  Using the $HOME variable will prevent it from failing when the users home directory does not live at the expected path.